fix options to yesno, move and rename composition
- question-type options is changed to yesno, to be more explicit
- compose the QuestionModal with Range or YesNo -content
